+++ Mods ChoGGi/Disable Drone Maintenance/Code/Script.lua
@@ -9,7 +9,7 @@ function OnMsg.ModsReloaded()
 	fire_once = true
 
 	-- version to version check with
-	local min_version = 37
+	local min_version = 38
 	local idx = table.find(ModsLoaded,"id","ChoGGi_Library")
 
 	-- if we can't find mod or mod is less then min_version (we skip steam since it updates automatically)
@@ -26,19 +26,21 @@ end
 local S
 local RetName
 local PopupToggle
+local RetAllOfClass
 
 -- generate is late enough that my library is loaded
 function OnMsg.ClassesGenerate()
 	S = ChoGGi.Strings
 	RetName = ChoGGi.ComFuncs.RetName
 	PopupToggle = ChoGGi.ComFuncs.PopupToggle
+	RetAllOfClass = ChoGGi.ComFuncs.RetAllOfClass
 end
 
 local orig_RequiresMaintenance_RequestMaintenance = RequiresMaintenance.RequestMaintenance
 -- only allow main if disable isn't
-function RequiresMaintenance:RequestMaintenance()
+function RequiresMaintenance:RequestMaintenance(...)
 	if not self.ChoGGi_DisableMaintenance then
-		orig_RequiresMaintenance_RequestMaintenance(self)
+		orig_RequiresMaintenance_RequestMaintenance(self,...)
 	end
 end
 
@@ -46,17 +48,13 @@ local function ToggleMain(obj)
 	if obj.ChoGGi_DisableMaintenance then
 		-- re-enable main
 		obj.ChoGGi_DisableMaintenance = nil
-		-- reset main requests (thanks mk-fg)
-		obj:AccumulateMaintenancePoints(0)
-		-- and check if building is malfunctioned then call a fix
-		if obj.accumulated_maintenance_points == obj.maintenance_threshold_current then
-			obj:RequestMaintenance()
-		end
+
+		obj:ResetMaintenanceState()
+		obj:RequestMaintenance()
 	else
 		-- disable it
 		obj.ChoGGi_DisableMaintenance = true
-		-- reset main requests (thanks mk-fg)
-		obj:ResetMaintenanceRequests()
+		obj:ResetMaintenanceRequests() --zero up reqs and interrupt drones
 	end
 end
 
@@ -106,7 +104,7 @@ function OnMsg.ClassesBuilt()
 						name = string.format([[Toggle maintenance on all %s.]],name),
 						hint = string.format([[Toggles maintenance on all %s (all will be set the same as this one).]],name),
 						clicked = function()
-							local objs = UICity.labels[context.class] or ""
+							local objs = RetAllOfClass(context.class)
 							for i = 1, #objs do
 								ToggleMain(objs[i])
 							end