+++ Mods ChoGGi/Mark Deposit Ground/metadata.lua
@@ -1,7 +1,7 @@
 return PlaceObj("ModDef", {
-	"title", "Mark Deposit Ground v0.3",
-	"version", 3,
-	"saved", 1542974400,
+	"title", "Mark Deposit Ground v0.4",
+	"version", 4,
+	"saved", 1543060800,
 	"image", "Preview.png",
 	"id", "ChoGGi_MarkDepositGround",
 	"steam_id", "1555446081",
@@ -10,10 +10,12 @@ return PlaceObj("ModDef", {
 	"code", {
 		"Code/Script.lua",
 		"Code/ModConfig.lua",
+		"Code/ShowConstruct.lua",
 	},
 	"description", [[Marks the ground around deposits so you can turn off the ugly signs and still see where they are.
 Marks are sized depending on max amount.
 
 Includes a ModConfig option to always hide the signs (you can still select the invisible sign and see the amount).
-And one to change all anomalies and or vista/research to the greenman model (suggestions for a different model?).]],
+An option to change all anomalies to the greenman model (suggestions for a different model?).
+And one more to show signs during construction mode.]],
 })